+++ ctwin32/ntdll.py
@@ -11,18 +11,16 @@ from collections import defaultdict as _defdict
 
 import ctypes
 from .wtypes import (
-    byte_buffer,
-    string_buffer,
     BOOLEAN,
     BYTE,
-    Struct,
-    Union,
+    byte_buffer,
     FILETIME,
+    GUID,
     HANDLE,
+    INT,
     LARGE_INTEGER,
     LONG,
     LONG_PTR,
-    INT,
     NTSTATUS,
     OSVERSIONINFOEX,
     PBOOLEAN,
@@ -32,12 +30,15 @@ from .wtypes import (
     PUNICODE_STRING,
     PVOID,
     SIZE_T,
+    string_buffer,
+    Struct,
     UINT_PTR,
     ULARGE_INTEGER,
     ULONG,
     ULONG_PTR,
-    UNICODE_STRING,
     UnicodeStrBuffer,
+    UNICODE_STRING,
+    Union,
     WCHAR,
     WCHAR_SIZE,
     WinError,
@@ -45,28 +46,30 @@ from .wtypes import (
     )
 from . import (
     ApiDll,
-    ref,
+    DIRECTORY_QUERY,
+    ERROR_NO_MORE_FILES,
+    GENERIC_READ,
     kernel,
     ns_from_struct,
-    suppress_winerr,
-    wtypes,
-    ERROR_NO_MORE_FILES,
-    DIRECTORY_QUERY,
-    GENERIC_READ,
-    SystemProcessInformation,
-    SystemProcessIdInformation,
-    SystemExtendedHandleInformation,
+    ProcessBasicInformation,
+    ProcessCommandLineInformation,
+    ProcessImageFileName,
+    ProcessWow64Information,
+    ref,
     STATUS_BUFFER_OVERFLOW,
     STATUS_BUFFER_TOO_SMALL,
     STATUS_INFO_LENGTH_MISMATCH,
     STATUS_MORE_ENTRIES,
     STATUS_NO_MORE_ENTRIES,
+    suppress_winerr,
+    SystemEnvironmentNameInformation,
+    SystemEnvironmentValueInformation,
+    SystemExtendedHandleInformation,
+    SystemProcessIdInformation,
+    SystemProcessInformation,
     ThreadBasicInformation,
     ThreadPriority,
-    ProcessCommandLineInformation,
-    ProcessImageFileName,
-    ProcessBasicInformation,
-    ProcessWow64Information,
+    wtypes,
     )
 
 _nt = ApiDll("ntdll.dll")
@@ -889,5 +892,66 @@ def NtQueryTimerResolution():
     cur = ULONG(0)
     raise_failed_status(_NtQueryTimerResolution(ref(min), ref(max), ref(cur)))
     return min.value, max.value, cur.value
+
+################################################################################
+
+class VARIABLE_NAME(Struct):
+    _fields_ = (
+        ("NextEntryOffset", ULONG),
+        ("VendorGuid", GUID),
+        ("Name", WCHAR * 1),
+        )
+
+class VARIABLE_NAME_AND_VALUE(Struct):
+    _fields_ = (
+        ("NextEntryOffset", ULONG),
+        ("ValueOffset", ULONG),
+        ("ValueLength", ULONG),
+        ("Attributes", ULONG),
+        ("VendorGuid", GUID),
+        ("Name", WCHAR * 1), # up to ValueOffset
+        # followed by ("Value", BYTE * ValueLength)
+        )
+
+_NtEnumerateSystemEnvironmentValuesEx = _nt.fun_fact(
+    "NtEnumerateSystemEnvironmentValuesEx",
+    (NTSTATUS, ULONG, PVOID, PULONG)
+    )
+
+def NtEnumerateSystemEnvironmentValuesEx(info_cls):
+    size = ULONG(0)
+    stat = _NtEnumerateSystemEnvironmentValuesEx(info_cls, None, ref(size))
+    if stat != STATUS_BUFFER_TOO_SMALL:
+        raise_failed_status(stat)
+
+    buf = byte_buffer(size.value)
+    raise_failed_status(
+        _NtEnumerateSystemEnvironmentValuesEx(info_cls, buf, ref(size))
+        )
+    addr = ctypes.addressof(buf)
+    offs = 0
+    next_offs = 1
+    result = []
+    if info_cls == SystemEnvironmentNameInformation:
+        noffs = VARIABLE_NAME.Name.offset
+        while next_offs:
+            vn = VARIABLE_NAME.from_buffer_copy(buf, offs)
+            name = ctypes.wstring_at(addr + offs + noffs)
+            result.append((vn.VendorGuid, name))
+            next_offs = vn.NextEntryOffset
+            offs += next_offs
+    elif info_cls == SystemEnvironmentValueInformation:
+        noffs = VARIABLE_NAME_AND_VALUE.Name.offset
+        while next_offs:
+            vnv = VARIABLE_NAME_AND_VALUE.from_buffer_copy(buf, offs)
+            name = ctypes.wstring_at(addr + offs + noffs)
+            val_addr = addr + offs + vnv.ValueOffset
+            value = ctypes.string_at(val_addr, vnv.ValueLength)
+            result.append((vnv.VendorGuid, name, vnv.Attributes, value))
+            next_offs = vnv.NextEntryOffset
+            offs += next_offs
+    else:
+        raise ValueError("unknown info class")
+    return result
 
 ################################################################################